requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

motor>=3.5.0

bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

pydantic[email]>=2.7.0

camoufox[geoip]>=0.4.0

playwright>=1.45.0

playwright-stealth>=1.0.6

python-slugify>=8.0.4

google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
httpx>=0.27.0

apscheduler>=3.10.4

python-dotenv>=1.0.1

cryptography>=42.0.0
